database/flask/src/requirements.txt
alembic==1.13.1
antiorm==1.2.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
db==0.1.1
db-sqlite3==0.0.1
flask==3.0.0
flask-login==0.6.3
flask-migrate==4.0.5
flask-sqlalchemy==3.1.1
idna==3.6
itsdangerous==2.1.2
jinja2==3.1.2
mako==1.3.0
markupsafe==2.1.3
numpy==1.26.2
packaging==23.2
pandas==2.1.4
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlalchemy==2.0.24
tzdata==2023.4
urllib3==2.1.0
werkzeug==3.0.1
gunicorn==20.1.0
openpyxl==3.0.9
drexel_jupyter_logger
xlrd==2.0.1
nbclient==0.7.4
nbconvert==7.13.1
nbformat==5.9.2


accessible-pygments==0.0.4
alabaster==0.7.13
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.0.1
click==8.1.7
comm==0.2.0
contourpy==1.2.0
correction-helper==2023.6
crashtest==0.4.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
docutils==0.18.1
dulwich==0.21.5
Engr131-Winter-2022==0.0.7
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.47.0
fqdn==1.5.1
friendly==0.7.21
friendly-styles==0.2
friendly-traceback==0.7.61
fsspec==2023.12.2
greenlet==3.0.3
html5lib==1.1
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.1
installer==0.7.0
ipykernel==6.28.0
ipython==8.19.0
ipywidgets==8.1.1
isoduration==20.11.0
jaraco.classes==3.2.3
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-book==0.15.1
jupyter-cache==0.6.1
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2